server_flask/requirements.txt
bleach==6.2.0
Flask==3.1.0
flask_cors==5.0.1
Flask_JWT_Extended==4.7.1
Flask_Login==0.6.3
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
python-dotenv==1.1.0
python_bcrypt==0.3.2
SQLAlchemy==2.0.40
